refactored code and displayed api data as a string in main
